+++ src/main/java/grit/stockIt/domain/order/service/OrderService.java
@@ -4,7 +4,6 @@ import grit.stockIt.domain.account.entity.Account;
 import grit.stockIt.domain.account.repository.AccountRepository;
 import grit.stockIt.domain.account.entity.AccountStock;
 import grit.stockIt.domain.account.repository.AccountStockRepository;
-import grit.stockIt.domain.matching.repository.RedisMarketDataRepository;
 import grit.stockIt.domain.matching.repository.RedisOrderBookRepository;
 import grit.stockIt.domain.order.dto.LimitOrderCreateRequest;
 import grit.stockIt.domain.order.dto.MarketOrderCreateRequest;
@@ -19,19 +18,14 @@ import grit.stockIt.domain.order.repository.OrderHoldRepository;
 import grit.stockIt.domain.order.repository.OrderRepository;
 import grit.stockIt.domain.stock.entity.Stock;
 import grit.stockIt.domain.stock.repository.StockRepository;
-import grit.stockIt.domain.stock.service.StockDetailService;
 import grit.stockIt.global.exception.BadRequestException;
-import grit.stockIt.global.exception.UntradeableStockException;
 import grit.stockIt.global.util.TransactionHandler;
 import grit.stockIt.global.websocket.manager.OrderSubscriptionCoordinator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.util.List;
 import java.util.Optional;
 
@@ -47,12 +41,8 @@ public class OrderService {
     private final OrderSubscriptionCoordinator orderSubscriptionCoordinator;
     private final OrderHoldRepository orderHoldRepository;
     private final AccountStockRepository accountStockRepository;
-    private final RedisMarketDataRepository redisMarketDataRepository;
-    private final StockDetailService stockDetailService;
     private final OrderAuthorizationService orderAuthorizationService;
-
-    @Value("${order.market.hold-buffer-rate:0.05}")
-    private BigDecimal marketHoldBufferRate;
+    private final OrderPricingService orderPricingService;
 
     // 지정가 주문 생성
     @Transactional
@@ -72,7 +62,7 @@ public class OrderService {
 
         // 매수 주문인 경우 거래 가능 종목인지 검증
         if (orderMethod == OrderMethod.BUY) {
-            validateStockTradeable(request.stockCode());
+            orderPricingService.validateStockTradeable(request.stockCode());
         }
 
         Order order = Order.createLimitOrder(
@@ -85,7 +75,7 @@ public class OrderService {
 
         BigDecimal holdAmount = BigDecimal.ZERO;
         if (orderMethod == OrderMethod.BUY) {
-            holdAmount = calculateHoldAmount(order); // 주문 금액 계산
+            holdAmount = orderPricingService.calculateHoldAmount(order); // 주문 금액 계산
             ensureSufficientCash(account, holdAmount); // 주문 가능 현금 확인
             account.increaseHoldAmount(holdAmount); // 홀딩 금액 증가
         } else if (orderMethod == OrderMethod.SELL) {
@@ -125,7 +115,7 @@ public class OrderService {
 
         // 매수 주문인 경우 거래 가능 종목인지 검증
         if (orderMethod == OrderMethod.BUY) {
-            validateStockTradeable(request.stockCode());
+            orderPricingService.validateStockTradeable(request.stockCode());
         }
 
         Order order = Order.createMarketOrder(
@@ -143,7 +133,7 @@ public class OrderService {
         if (orderMethod == OrderMethod.SELL) {
             applySellHold(order);
         } else if (orderMethod == OrderMethod.BUY) {
-            holdAmount = calculateMarketHoldAmount(stock.getCode(), order.getQuantity());
+            holdAmount = orderPricingService.calculateMarketHoldAmount(stock.getCode(), order.getQuantity());
             ensureSufficientCash(account, holdAmount);
             account.increaseHoldAmount(holdAmount);
         }
@@ -245,43 +235,6 @@ public class OrderService {
         return new PendingOrdersResponse(orderItems);
     }
 
-    private BigDecimal calculateHoldAmount(Order order) {
-        return order.getPrice().multiply(BigDecimal.valueOf(order.getRemainingQuantity()));
-    }
-
-    // 시장가 주문 홀딩 금액 계산
-    private BigDecimal calculateMarketHoldAmount(String stockCode, int quantity) {
-        // 1. Redis 캐시에서 먼저 조회
-        BigDecimal lastPrice = redisMarketDataRepository.getLastPrice(stockCode)
-                .orElseGet(() -> {
-                    // 2. 캐시에 없으면 KIS API 호출
-                    log.info("캐시에 현재가가 없어 KIS API 호출: stockCode={}", stockCode);
-                    try {
-                        BigDecimal price = stockDetailService.getCurrentPrice(stockCode)
-                                .block(java.time.Duration.ofSeconds(5));
-                        if (price == null || price.signum() <= 0) {
-                            throw new BadRequestException("KIS API에서 현재가를 가져올 수 없습니다.");
-                        }
-                        // KIS API 결과는 StockDetailService에서 이미 Redis에 저장됨
-                        return price;
-                    } catch (Exception e) {
-                        log.error("KIS API 현재가 조회 실패: stockCode={}", stockCode, e);
-                        throw new BadRequestException("최근 체결가 정보를 찾을 수 없습니다.");
-                    }
-                });
-        
-        if (lastPrice.signum() <= 0) {
-            throw new BadRequestException("최근 체결가가 유효하지 않습니다.");
-        }
-        BigDecimal bufferRate = Optional.ofNullable(marketHoldBufferRate).orElse(BigDecimal.valueOf(0.05));
-        if (bufferRate.signum() < 0) {
-            bufferRate = BigDecimal.ZERO;
-        }
-        BigDecimal bufferFactor = BigDecimal.ONE.add(bufferRate);
-        BigDecimal baseAmount = lastPrice.multiply(BigDecimal.valueOf(quantity));
-        return baseAmount.multiply(bufferFactor).setScale(2, RoundingMode.UP);
-    }
-
     private void ensureSufficientCash(Account account, BigDecimal holdAmount) {
         if (account.getAvailableCash().compareTo(holdAmount) < 0) {
             throw new BadRequestException("주문 가능 현금이 부족합니다.");
@@ -332,24 +285,4 @@ public class OrderService {
         }
     }
 
-    // 거래 가능 종목인지 검증
-    private void validateStockTradeable(String stockCode) {
-        try {
-            var stockDetail = stockDetailService.getStockDetail(stockCode)
-                    .block(java.time.Duration.ofSeconds(5));
-            
-            if (stockDetail == null || !Boolean.TRUE.equals(stockDetail.tradeable())) {
-                String reason = stockDetail != null && stockDetail.untradeableReason() != null
-                        ? stockDetail.untradeableReason()
-                        : "이 종목은 AI 분석이 불가능하여 거래가 제한됩니다.";
-                throw new UntradeableStockException(reason);
-            }
-        } catch (UntradeableStockException e) {
-            throw e;
-        } catch (Exception e) {
-            log.error("종목 거래 가능 여부 확인 실패: stockCode={}", stockCode, e);
-            throw new UntradeableStockException("종목 거래 가능 여부를 확인할 수 없습니다.");
-        }
-    }
 }
-
